+++ fluent_blogs/utils/compat.py
@@ -4,18 +4,20 @@ Django compatibility features
 from django.conf import settings
 
 __all__ = (
-    'now', 'get_user_model', 'get_user_model_name',
+    'now', 'utc',
+    'get_user_model', 'get_user_model_name',
     'patterns', 'url', 'include',
 )
 
 
 # The timezone support was introduced in Django 1.4, fallback to standard library for 1.3.
 try:
-    from django.utils.timezone import now
+    from django.utils.timezone import now, utc
 except ImportError:
     # Django < 1.4
     from datetime import datetime
     now = datetime.now
+    utc = None   # very naive
 
 
 # Support for custom User models in Django 1.5+